knowledge_panels/requirements.txt
markdown==3.4.4
openfoodfacts==0.1.10
pydantic==2.4.2
ruamel.yaml==0.17.35